backend/requirements.txt
fastapi
uvicorn[standard]
httpx
anthropic
gspread
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
